Version 1.1
Removed Local Caching (including the CacheUtility class).

Local Caching was removed because it was unless. It was more of a slow down than anything. Performance was increased when I switched
VatsimSync (Discord) to version 1.1 of the VatsimMemberAPI.

NOTE: You will not need to change any of your code if you are only using the Client Builder (which is how you should be using it).